requirements.aging.txt
numpy>=1.18.0,<1.20.0
opencv-python>=4.1.2.0
torch==1.2.0
torchvision==0.4.0
pyyaml==5.4.1
Pillow==8.2.0
tqdm==4.59.0
azureml-core==1.27.0
azureml-pipeline==1.27.0
azureml-dataset-runtime[fuse]==1.27.0
azureml-defaults==1.27.0
azureml-contrib-services==1.27.0